[frontend] Add password input to user create form
